+++ molo/surveys/urls.py
@@ -1,7 +1,8 @@
 
 from django.conf.urls import url
 from molo.surveys.views import (
-    SurveySuccess, submission_article, get_segment_user_count
+    SurveySuccess, ResultsPercentagesJson, submission_article,
+    get_segment_user_count
 )
 
 
@@ -10,6 +11,11 @@ urlpatterns = [
         r"^(?P<slug>[\w-]+)/success/$",
         SurveySuccess.as_view(),
         name="success"
+    ),
+    url(
+        r"^(?P<slug>[\w-]+)/results_json/$",
+        ResultsPercentagesJson.as_view(),
+        name="results_json"
     ),
     url(
         r'^submissions/(\d+)/article/(\d+)/$',